+++ src/uwsgiconf/settings.py
@@ -6,6 +6,7 @@ ENV_CONF_ALIAS = 'UWSGICONF_CONF_ALIAS'
 ENV_CONF_READY = 'UWSGICONF_READY'
 ENV_FORCE_STUB = 'UWSGICONF_FORCE_STUB'
 ENV_MAINTENANCE = 'UWSGICONF_MAINTENANCE'
+ENV_SKIP_TASK = 'UWSGICONF_SKIP_TASK_{task_name}'
 ENV_MAINTENANCE_INPLACE = 'UWSGICONF_MAINTENANCE_INPLACE'
 
 
@@ -26,4 +27,14 @@ def get_maintenance_inplace() -> bool:
     Introduced as a function to support embedded mode.
 
     """
-    return int(environ.get(ENV_MAINTENANCE_INPLACE, 0)) != 0
+    return environ.get(ENV_MAINTENANCE_INPLACE, "0") != "0"
+
+
+def get_skip_task(task_name: str, *, env_var: str = ENV_SKIP_TASK) -> bool:
+    """Returns a flag from env indicating whether a task should be skipped.
+
+    :param task_name: Task name.
+    :param env_var: Environment variable to check.
+        Note: if contains `{task_name}` placeholder, then it'll be replaced with the value from `name`.
+    """
+    return environ.get(env_var.format(task_name=task_name.upper()), "0") != "0"